Handle legacy /logout
